Move some exceptions to its own package (making them not part of public API)
