+++ config-model/src/main/java/com/yahoo/vespa/model/VespaModelFactory.java
@@ -20,7 +20,7 @@ import com.yahoo.config.model.api.ValidationParameters;
 import com.yahoo.config.model.application.provider.ApplicationPackageXmlFilesValidator;
 import com.yahoo.config.model.builder.xml.ConfigModelBuilder;
 import com.yahoo.config.model.deploy.DeployState;
-import com.yahoo.config.provision.TransientException;
+import com.yahoo.config.provision.exception.TransientException;
 import com.yahoo.config.provision.Zone;
 import com.yahoo.vespa.config.VespaVersion;
 import com.yahoo.vespa.model.application.validation.Validation;
