+++ config-provisioning/src/main/java/com/yahoo/config/provision/exception/TransientException.java
@@ -1,5 +1,5 @@
 // Copyright 2019 Oath Inc. Licensed under the terms of the Apache 2.0 license. See LICENSE in the project root.
-package com.yahoo.config.provision;
+package com.yahoo.config.provision.exception;
 
 /**
  * A provisioning exception that is considered transient. Exceptions that we expect to recover from after a short
