+++ node-repository/src/main/java/com/yahoo/vespa/hosted/provision/lb/LoadBalancerServiceException.java
@@ -1,7 +1,7 @@
 // Copyright 2019 Oath Inc. Licensed under the terms of the Apache 2.0 license. See LICENSE in the project root.
 package com.yahoo.vespa.hosted.provision.lb;
 
-import com.yahoo.config.provision.TransientException;
+import com.yahoo.config.provision.exception.TransientException;
 
 /**
  * Transient exception thrown on behalf of a {@link LoadBalancerService}.
