+++ node-repository/src/main/java/com/yahoo/vespa/hosted/provision/provisioning/Activator.java
@@ -5,7 +5,7 @@ import com.yahoo.config.provision.ApplicationId;
 import com.yahoo.config.provision.ClusterMembership;
 import com.yahoo.config.provision.ClusterSpec;
 import com.yahoo.config.provision.HostSpec;
-import com.yahoo.config.provision.ParentHostUnavailableException;
+import com.yahoo.config.provision.exception.ParentHostUnavailableException;
 import com.yahoo.transaction.Mutex;
 import com.yahoo.transaction.NestedTransaction;
 import com.yahoo.vespa.hosted.provision.Node;
